Move App to app directory
